// app/api/models/responseAcademixDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * My School APIs
 * My School API documentation
 * OpenAPI spec version: 1.0
 */
import type { MessageDto } from './messageDto';
import type { MetadataDto } from './metadataDto';

export interface ResponseAcademixDto {
  data: MessageDto;
  message: string;
  status: boolean;
  statusCode: number;
  metadata: MetadataDto;
}